Bind to temp_f

+++ WeatherApp.WebUI/Models/WeatherReport.cs
@@ -5,5 +5,5 @@ namespace WeatherApp.WebUI.Models;
 public class WeatherReport
 {
     [JsonPropertyName("temp_f")]
-    public int Fahrenheit { get; set; }
+    public double Fahrenheit { get; set; }
 }
